Create codeExample.py
Sample of a much larger project.

This script in particular is used to optimize models for Team Win % Predictions when given starting player performance for the season (players who played at least half of the expected playing time). 

This is ultimately used to project team performance from predicted starting player performance for the season. 

To demonstrate my thought process and software design style, I have not editted or reviewed this code any further since its completion, I am not aware of any bugs. However, I have redacted all CSV file names, as those contain outputs from other scripts of player performance evaluations. The entire system is not completed yet (but this script is), so I have not made it public as it is a WIP.